+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.30.0] - 2025-11-06
+
 ### Changed
 
 - `ParamsBiotSavart`: change default value of `quadrature_near_singularity`. Now it is equal